Call reusable action for go version updates (#204)

+++ pkg/logger/logger.go
@@ -58,7 +58,7 @@ func New(level int, formatter string) {
 	logrusInstance.SetLevel(logrus.Level(defaultLevel))
 
 	if level >= minLevel || level <= maxLevel {
-		logrusInstance.SetLevel(logrus.Level(level))
+		logrusInstance.SetLevel(logrus.Level(uint32(level))) // #nosec G115
 	}
 
 	switch formatter {
